refactor: Reorganize utils into a utils folder in src
