+++ src/utils/formatColor.ts
